requirements.txt
gwpy>=3.0.0
gwosc>=0.7.0
pandas>=2.1
Pillow>=10.0
matplotlib>=3.8
tqdm>=4.66
requests>=2.31
scikit-learn>=1.4
h5py>=3.10
joblib>=1.3
ligo-segments
numpy
